refactor(bench): split run.rs god-file into modules
